Ran analyze and patched compile error
added todo’s for places were the proper fix for the compile error needs
more investigation. two types: missing return statement and comparing
byte to null.

+++ FtcRobotController/src/main/java/com/qualcomm/hardware/MatrixServoController.java
@@ -72,7 +72,7 @@ public class MatrixServoController implements ServoController {
 
     private void m50a(int i) {
         if (i < 1 || i > 4) {
-            throw new IllegalArgumentException(String.format("Channel %d is invalid; valid channels are 1..%d", new Object[]{Integer.valueOf(i), Byte.valueOf((byte) 4)}));
+            throw new IllegalArgumentException(String.format("Channel %d is invalid; valid channels are 1..%d", new Object[]{i, (byte) 4}));
         }
     }
